Fixed paleofetch on M1 * chaged vm_stat code, so it shows almost correct memory usage * bitshifted where I could * added CC to makefile

+++ src/macintosh.h
@@ -10,4 +10,4 @@ static char *get_ram_usage(),
             *get_resolution(),
             *get_gpu();
 
-static uint get_mem_from_vm_stat();
+static uint64_t get_mem_from_vm_stat();
